Make old tests pass

+++ tsplp/include/DependencyHelpers.hpp
@@ -6,7 +6,7 @@
 
 namespace tsplp
 {
-xt::xtensor<int, 2> CreateTransitiveDependencies(xt::xtensor<int, 2> weights);
+xt::xtensor<double, 2> CreateTransitiveDependencies(xt::xtensor<double, 2> weights);
 
 class DependencyGraph
 {
@@ -18,10 +18,10 @@ private:
     std::vector<std::pair<size_t, size_t>> m_node2incomingSpanMap;
     std::vector<std::pair<size_t, size_t>> m_node2outgoingSpanMap;
 
-    const xt::xtensor<int, 2>& m_weights;
+    const xt::xtensor<double, 2>& m_weights;
 
 public:
-    explicit DependencyGraph(const xt::xtensor<int, 2>& weights);
+    explicit DependencyGraph(const xt::xtensor<double, 2>& weights);
 
     [[nodiscard]] const auto& GetArcs() const { return m_arcs; }
     [[nodiscard]] std::span<const size_t> GetIncomingSpan(size_t n) const;
